config: make empty Config behavior default for Invocables (#202)
This PR refactors the `config_cls` method from `Invocable` so that it is no longer
abstract. This will provide a default behavior for plugins and packages, eliminating
boilerplate when configuration is not needed.

Packages and plugins can still override the behavior by defining their own `config_cls`
implementation.

If this leads to issues with users not properly providing configuration due to API mishaps,
we can add a Configurable ABC for use with invocables. This might work like:

```
class Configurable(ABC):
  @abstractmethod
  def config_cls(self) -> Type[Config]:
    raise NotImplementedError()

class MyPackage(Configurable, Invocable):

  class MyPackageConfig(Config):
    foo: str

  def config_cls(self) -> Type[Config]:
    return MyPackageConfig
```