fix: ensure expectation is met in lookups/creates (#303)
Steamship should be a good 🚢 . A good ship should not lie to users. It
would be a lie to tell a user that a File was created when the object
returned is not a File. Hence, Steamship must enforce expectations.

Without this PR, it is possible, though rare, for the following to
happen:

```
f = File.create(client=client, blocks=[Block(text="something filtered by WAF")])
print(f.handle)
```

```
steamship.base.error.SteamshipError: ... | 'str' object has no attribute 'handle'
```

Co-authored-by: Douglas Reid <doug@steamship.com>