feat: add 'ship use' command for easy instance creation
This adds a basic command to the CLI to allow for rapid instance creation.
It must be run from a directory containing a Steamship manifest file.

Configuration may be passed as a JSON string, or as a path to JSON file.

NOTE: At the moment, this must be run after 'ship deploy', as no attempt is
made to create a package/plugin before creation of an instance. This could
be addressed, if desired.