On generator return, only rewrite blocks if the result has data (#352)
Currently, it's not possible to write an async generator plugin, because
the shim code in generator.py attempts to rewrite data. In the async
case, where data is not yet filled, this will throw an error.

This PR adds a condition so that the blocks are only rewritten if data
is present.