refactor: cleanup agent class objects (#384)
This PR represents an attempt to cleanup, document, and rationalize the
existing set of classes in `agent` hierarchy. I also wanted to remove as
much vestigial and unused surface (primarily related to async execution)
in an attempt to simplify and reduce surface area for implementors.

As part of this PR, I updated the example assistant to show how
synchronous agent responses could be done using `emitFns`. This then led
to an update of the REPL to allow flexible identification of
`AgentService` methods. I think we need more thought there.

Because we've introduced an `LLM` class, I took the `get_llm()` method
that created ad-hoc ones from context off of the context and added
helpers to create contexts with LLM instances. This was an attempt to
rationalize usage within Tools, etc. I'm open to other approaches there.

---------

Co-authored-by: Douglas Reid <doug@steamship.com>
Co-authored-by: Ted Benson <ted@steamship.com>