Package Mixins (#396)
This PR builds on the dynamic route registration of #393 and sketches
the concept of an explicit Package Mixin, which is added to a
PackageService in its init method, and adds routes to the package.

The goal is that things like Telegram Transport and Ask My X could be
mixed into an agent without having to resort to multiple inheritance.

Right now, the mixins get to see the PackageService's config during
instance_init. This would mean that the PackageService's config would
need to include the union of parameters for itself and all its mixins. I
can't yet think of a better way.

---------

Co-authored-by: Ted Benson <ted@steamship.com>