Emit Bug Fix [explanation required] (#409)
This is a fix for a bug where the first message to a web bot works, but
subsequent calls do not.

The issue appears to be that SOMEHOW, the emit_func from the previous
run is still present in memory. This does not make any sense to me, nor
does the "existing emit functions" logging support that.

AND YET - The other logging, of which emit function is being used,
supports the existing-functions theory. Before the change to overwrite
the list, if you sent a message to the Web client, then a message to
Telegram, the telegram call would attempt to use the web client's emit.
Also, changing the code to always overwrite the emit function list fixes
the issue.