Transports as Mixins (#411)
This PR refactors Transports to be Package Mixins. This allows them to
be additive aspects to a an Agent Package without requiring subclassing
or conflicting with one another.

- Web Widget and Telegram transports promoted from Experimental to
agents package.
- Functionality from TelegramAgent and SteamshipWeb agent merged into
the transports to make them mixins.
- Old versions of Bots, Agents, and Transports removed to reduce surface
area.
- Introduce mock Telegram service and first actual SDK tests for web
widget to hopefully reduce regression issues.

Other intentionally removed functionality:

- Transport as a mixin superclass can't provide any routes, because they
would always conflict as soon as you had multiple Transports. This
removes the prototype `deinit` functionality. This could / should be
readded with Engine support so that there is a guarantee it is called.