feat(files): add helper util for uploading a local file (#402)
Some users are struggling with the wrapper code to upload local files
into Steamship. This happens in replit, as well as in other dev
environments.

This PR attempts to provide a helper utility to simplify the cognitive
load on users. The use case this addresses is the "I just want to upload
this file" case.

Expected usage:
```python
steamship_file = File.from_local(client, "our_business_plan.txt")
```

---------

Co-authored-by: Douglas Reid <doug@steamship.com>