feat(functions): add support for mapping tools to openai functions (#435)
In order to build `Agent`s that can exploit OpenAI's `function_call`
interface, we will likely want a way to directly map Steamship Tools to
OpenAI function objects.

This PR proposes a way to accomplish this task. It is a parallel PR to
https://github.com/steamship-plugins/gpt4/pull/8 (which validates
~equivalent input will work with OpenAI prompting).

With the default scheme, any tool that accepts as input either:
- a single text string
- a single non-text block
- a text string and a non-text block

should be usable in OpenAI `function_call` based prompting.

More work will be necessary to support multiple inputs of the same type
(text or uuid). Submitting parameters of type `array` was causing errors
returned by OpenAI, so this may be an unimplemented feature on their end
that we need to work around.

No attempt here was made to add built-in Tag support for tags need for
`name` tags and `function` messages. That should be done in a subsequent
PR.

Co-authored-by: Douglas Reid <doug@steamship.com>