feat: Standardize AgentREPL handling of output; provide default `prompt` (#458)
This PR attempts to simplify the complexity of developing & creating an
Agent in two ways:

1) It standardizes the handling of Tool/Agent output within the Repl
(previously Tool output was handled by the REPL, but Agent output was
handled by the Agent).

2) It provides a default `prompt` method if none is provided during REPL
instantiation. This permits the 90% case of agent development to be
nothing more than the agent definition itself, without the `prompt`
method which sets up the debugging harness.