fix(tests): add retries to model checkpoint test (#486)
There is a test flake around DEFAULT checkpoints when multiple test runs
happen at the same time. This PR is an attempt to solve that by
attempting to rerun the checks until the hashes match.

I don't know if that's the best way to handle this issue, but it was all
I could think of at the moment. Someone with a deeper understanding of
the intended behavior here could possibly come up with a better
solution.

---------

Co-authored-by: Douglas Reid <doug@steamship.com>