feat(agents): add caching for LLM calls and Actions (#493)
This is an initial PR for adding caching functionality to Agents. It is
aimed at reducing cost and latency for repeated requests.

It attempts to provide two levels of caching:
(1) prompts to the LLMs (prompt-to-Action cache) and (2) action
executions (Action-to-outputs cache).

Cache behavior can be dynamically control per call to `prompt`.

NOTE: This PR is NOT introducing semantic caches (text/block uuids must
match precisely).
FOLLOW-ONS: Allow Tool-default opt-in/opt-out of caching

---------

Co-authored-by: Douglas Reid <doug@steamship.com>