feat: Slack Transport for Agent <> Slack communication (#460)
@dkolas this is the Slack transport that I wrote the other day. It
works, but it's not polished and needs a bit of touch-up.

Before I do that, I was hoping for a quick scan for any high level
design disagreements within so that I can fix those before polishing off
the edges.

My thinking was to get something relatively simple in, along with
documentation, and then we'll learn where the gremlins lie (as we did
with Telegram) once people start to use it.

Also hoping it'll serve useful for us to consider a broader set of
"Transports" together vis-a-vis developer/user UX issues ---- e.g. is
there a universal UI/UX that manages transports? Are they ad hoc? Etc.

---------

Co-authored-by: Enias Cailliau <hello@enias.ai>
Co-authored-by: Enias Cailliau <enias@steamship.com>