feat(streaming): proof of concept for streaming agentservice execution
This PR presents a series of changes that should support a way to
stream response information back to a client via an AgentService.

In order to achieve the streaming result, a new method on the AgentService
is exposed: `async_prompt`. This new method returns a new `StreamingResponse`
that has two fields: `task` and `file`. These fields provide access to (a)
the async task that will be streaming results and (b) a file (here the
`ChatHistory` file) to which all status messages and assistant interactions
will be saved.

This PR relies on a full deployment of https://github.com/steamship-plugins/gpt4/pull/10
to the target environment for testing / validation.