feat: `ship run local` should direct to, and work with, the live Agent Management Console (#561)
Before, the `ship run local` command directed to a limited debugging
page which had just a bare chat interface that hit a local endpoint
provided by URL query parameter.

But this left several important things unaddressed:
- How should the user connect Telegram or Slack to their locally running
agent?
- How could the user see the files, tasks, and logs being created by
their locally running agent?
- How could the user try out multiple chat sessions?

This PR connects the locally running agent to the live console. It
removes the code that simulated the "proxy responses" from the local
server so that it can communicate correctly via the Steamship Proxy in
the wild.


https://github.com/steamship-core/python-client/assets/63262/87e9aa50-44b5-4f7b-be3e-3cb1fc67e576