fix(agents): generate proper history blocks in agent (#570)
With the inflight changes for streaming, we unfortunately merged a
commit that left message selection in a broken state. The existing
testing was not enough to capture the issue. This PR attempts to restore
proper message selection functionality.

Follow on PRs will be necessary to clean up and streamline the selection
bits added in this PR.

Co-authored-by: Douglas Reid <doug@steamship.com>