feat(streaming): initial implementation of streaming agentservice execution (#554)
This PR establishes an initial implementation of streaming utilities for
`AgentService` endpoints, based on `FunctionsBasedAgent` runs. This will
allow clients to invoke an agent run through an async endpoint and
stream, via SSE, block creation events related to the agent's execution.
These events can include status messages (from Agents and Tools) as well
as generated blocks (from LLMs and Tools).

With this code, new async endpoints can be exposed with code like the
following:

```python
    @post("async_prompt")
    def async_prompt(
        self, prompt: Optional[str] = None, context_id: Optional[str] = None, **kwargs
    ) -> StreamingResponse:
        ctx_id, history_file = self._streaming_context_id_and_file(context_id=context_id, **kwargs)
        task = self.invoke_later(
            "/prompt", arguments={"prompt": prompt, "context_id": ctx_id, **kwargs}
        )
        return StreamingResponse(task=task, file=history_file)
```

The PR ensures all `Block`s created in the service of a `run_agent` call
are tagged with the proper `request-id`, and that all calls to the LLM
are streaming-compatible.

A test is provided that demonstrates an approach to consuming a stream
of events, based on the sseclient-py library. A Generator is constructed
that emits `Block`s until a terminal block for a request is found in the
stream.

---------

Co-authored-by: Douglas Reid <doug@steamship.com>